// src/Polyglot.Frontend/src/app/api/model/mcpTransportMode.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const McpTransportMode = {
    Auto: 'Auto',
    StreamableHttp: 'StreamableHttp',
    Sse: 'Sse'
} as const;
export type McpTransportMode = typeof McpTransportMode[keyof typeof McpTransportMode];