// src/Polyglot.Frontend/src/app/api/model/mcpServerDto.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { McpTransportMode } from './mcpTransportMode';


export interface McpServerDto { 
    id: string;
    name: string;
    url: string;
    transportMode: McpTransportMode;
    enabled: boolean;
    isGlobal: boolean;
    hasAuthHeader: boolean;
    canManage: boolean;
}
export namespace McpServerDto {
}